requirements-hf.txt
gradio==4.44.0
Pillow==10.4.0
requests==2.32.4
transformers==4.44.2
torch==2.4.1
diffusers==0.30.2

